MedicalQA_FlyAI_self/requirements.txt
flyai-gpu
six
jieba
numpy
regex
ufile
requests
tornado
scikit-learn
tensorflow-gpu==1.14.0
tensorflow-hub
keras==2.2.5